+++ crates/router/src/configs/secrets_transformers.rs
@@ -572,6 +572,7 @@ pub(crate) async fn fetch_raw_secrets(
         zero_mandates: conf.zero_mandates,
         network_transaction_id_supported_connectors: conf
             .network_transaction_id_supported_connectors,
+        card_only_mit_supported_connectors: conf.card_only_mit_supported_connectors,
         list_dispute_supported_connectors: conf.list_dispute_supported_connectors,
         required_fields: conf.required_fields,
         delayed_session_response: conf.delayed_session_response,