+++ crates/router/src/core/payments/flows/authorize_flow.rs
@@ -1226,6 +1226,7 @@ fn transform_response_for_pre_authenticate_flow(
                 network_txn_id,
                 connector_response_reference_id,
                 incremental_authorization_allowed,
+                authentication_data,
                 charges,
             },
         ) => {
@@ -1249,6 +1250,7 @@ fn transform_response_for_pre_authenticate_flow(
                     network_txn_id,
                     connector_response_reference_id,
                     incremental_authorization_allowed,
+                    authentication_data,
                     charges,
                 },
             )